update FilterUrlLink function
