+++ sites/public/pages/eligibility/household.tsx
@@ -19,6 +19,7 @@ import { ELIGIBILITY_SECTIONS } from "../../lib/constants"
 import { EligibilityContext } from "../../lib/EligibilityContext"
 import FormBackLink from "../../src/forms/applications/FormBackLink"
 import { eligibilityRoute } from "../../lib/helpers"
+import { getFilterUrlLink } from "./filterUrlLink"
 
 const EligibilityHouseholdSize = () => {
   const router = useRouter()
@@ -43,6 +44,13 @@ const EligibilityHouseholdSize = () => {
 
   if (eligibilityRequirements.completedSections <= CURRENT_PAGE) {
     eligibilityRequirements.setCompletedSections(CURRENT_PAGE + 1)
+  }
+
+  const onClick = () => {
+    const data = getValues()
+    const { householdSize } = data
+    eligibilityRequirements.setHouseholdSizeCount(householdSize)
+    void router.push(getFilterUrlLink(eligibilityRequirements))
   }
 
   const householdSizeRanges = ["one", "two", "three", "four", "five", "six", "seven", "eight"]
@@ -87,7 +95,11 @@ const EligibilityHouseholdSize = () => {
               <Button className="mx-2 mt-6" styleType={AppearanceStyleType.primary}>
                 {t("t.next")}
               </Button>
-              <Button className="mx-2 mt-6" styleType={AppearanceStyleType.primary}>
+              <Button
+                onClick={handleSubmit(onClick)}
+                className="mx-2 mt-6"
+                styleType={AppearanceStyleType.primary}
+              >
                 {t("t.viewListings")}
               </Button>
             </div>
