+++ sites/public/pages/listings.tsx
@@ -183,8 +183,8 @@ const ListingsPage = () => {
   // eslint-disable-next-line @typescript-eslint/unbound-method
   const { handleSubmit, register, errors } = useForm()
   const onSubmit = (data: ListingFilterState) => {
-    if (data[FrontendListingFilterStateKeys.includeNulls] === false) {
-      delete data[FrontendListingFilterStateKeys.includeNulls]
+    if (data?.[FrontendListingFilterStateKeys.includeNulls] === false) {
+      delete data?.[FrontendListingFilterStateKeys.includeNulls]
     }
     setFilterModalVisible(false)
     setQueryString(/*page=*/ 1, data)
